requirements.txt
Django==3.2.9
python-decouple==3.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
